test(seat): 트랜잭션-락 분리 구조에 맞춰 테스트 재작성
- SeatHoldServic/SeatAssignmentService는 validation, 락 획득/해제, 위임 호출만 검증하도록 변경.
- SeatHoldTransactionalService/SeatAssignmentTransactionalService에 비즈니스 로직 테스트(만료 연장, 좌석 교체, 충돌 예외 등)를 분리하여 작성.
- 예외 발생 시에도 락이 반드시 해제되는지 검증하는 테스트 추가.
